# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a production-ready payment gateway service that demonstrates enterprise-grade architecture with idempotent operations, asynchronous processing, and reliable message delivery.",
        "title": "Payment Gateway API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/payments": {
            "post": {
                "description": "Create a new payment with built-in idempotency protection. If a payment with the same idempotency_key already exists, it returns the existing payment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create payment",
                "parameters": [
                    {
                        "description": "Payment creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Existing payment returned (idempotent)",
                        "schema": {
                            "$ref": "#/definitions/response.CreatePaymentResponse"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CreatePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/{id}": {
            "get": {
                "description": "Retrieve detailed information about a payment including its current status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get payment details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns basic health information about the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Service health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreatePaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Payment amount (must be greater than 0)\nrequired: true\nexample: 100.50",
                    "type": "number"
                },
                "currency": {
                    "description": "Payment currency (ETB or USD)\nrequired: true\nenum: ETB,USD\nexample: USD",
                    "type": "string"
                },
                "idempotency_key": {
                    "description": "Unique idempotency key to prevent duplicate payments\nrequired: true\nexample: order-12345",
                    "type": "string"
                }
            }
        },
        "entities.PaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusSuccess",
                "PaymentStatusFailed"
            ]
        },
        "response.CreatePaymentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Payment creation timestamp\nexample: 2024-01-05T10:30:00Z",
                    "type": "string"
                },
                "payment_id": {
                    "description": "Unique payment identifier\nexample: 550e8400-e29b-41d4-a716-446655440000",
                    "type": "string"
                },
                "status": {
                    "description": "Current payment status\nenum: PENDING,SUCCESS,FAILED\nexample: PENDING",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.PaymentStatus"
                        }
                    ]
                }
            }
        },
        "response.DetailedErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FieldError"
                    }
                },
                "ft_reference": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "response.FieldError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.GetPaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Payment amount as string\nexample: 100.50",
                    "type": "string"
                },
                "created_at": {
                    "description": "Payment creation timestamp\nexample: 2024-01-05T10:30:00Z",
                    "type": "string"
                },
                "currency": {
                    "description": "Payment currency\nenum: ETB,USD\nexample: USD",
                    "type": "string"
                },
                "id": {
                    "description": "Unique payment identifier\nexample: 550e8400-e29b-41d4-a716-446655440000",
                    "type": "string"
                },
                "idempotency_key": {
                    "description": "Idempotency key used for the payment\nexample: order-12345",
                    "type": "string"
                },
                "processed_at": {
                    "description": "Payment processing completion timestamp (null if still pending)\nexample: 2024-01-05T10:30:05Z",
                    "type": "string"
                },
                "status": {
                    "description": "Current payment status\nenum: PENDING,SUCCESS,FAILED\nexample: SUCCESS",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.PaymentStatus"
                        }
                    ]
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.DetailedErrorResponse"
                },
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}